Выносит отрисовку точки в отдельный класс
