+++ src/presenter/app-presenter.js
@@ -3,12 +3,10 @@ import InfoView from '../view/header-views/info-view';
 import FiltersView from '../view/header-views/filters-view';
 import EventsSortView from '../view/main-views/sort-view';
 import EventsListView from '../view/main-views/list-view';
-import PointView from '../view/main-views/point-view';
-import EditPointView from '../view/main-views/edit-point-view';
-import { getRandomArrayElement, isEscapeKey } from '../utils/utils';
 import EventsMessage from '../view/main-views/message-view';
 import { TYPES_OF_SORT } from '../utils/const';
 import { generateFilter } from '../mock/filter';
+import PointPresenter from './point-presenter';
 
 export default class EventsPresenter {
   #eventsElement = null;
@@ -24,6 +22,7 @@ export default class EventsPresenter {
   #eventMessageElement = new EventsMessage();
   #eventsInfoElement = new InfoView();
 
+
   constructor(eventsElement, mainElement, filtersElement, pointModel) {
     this.#eventsElement = eventsElement;
     this.#mainElement = mainElement;
@@ -38,67 +37,21 @@ export default class EventsPresenter {
 
     this.#renderInfo();
     this.#renderFilter(this.#points);
-    this.#renderPoints(this.#points);
-  }
-
-  #renderPoint(point, destination, offers) {
-    const pointElement = new PointView(
-      point,
-      destination,
-      offers,
-      {
-        onButtonClick: () => {
-          replacePointToForm.call(this);
-          document.addEventListener('keydown', onEscKeydown);
-        }
-      });
-    const pointEditElement = new EditPointView(point,
-      destination,
-      offers,
-      {
-        onButtonClick: () => {
-          replacePointToCard.call(this);
-          document.removeEventListener('keydown', onEscKeydown);
-        },
-        onFormSubmit: (evt) => {
-          evt.preventDefault();
-          replacePointToCard.call(this);
-          document.removeEventListener('keydown', onEscKeydown);
-        }
-      });
-
-    function replacePointToForm() {
-      this.#eventsListElement.element.replaceChild(pointEditElement.element, pointElement.element);
-    }
-
-    function replacePointToCard() {
-      this.#eventsListElement.element.replaceChild(pointElement.element, pointEditElement.element);
-    }
-
-    function onEscKeydown(evt) {
-      if(isEscapeKey(evt)) {
-        evt.preventDefault();
-        replacePointToCard.call(this);
-        document.removeEventListener('keydown', onEscKeydown);
-      }
-    }
-
-    render(pointElement, this.#eventsListElement.element);
+    this.#renderSort(TYPES_OF_SORT);
+    this.#renderList(this.#points);
   }
 
   #renderPoints(points) {
-    const randomPoint = () => getRandomArrayElement(points);
     const getDestination = (point) => this.#destinations.find((item) => item.id === point.destination ? item.id === point.destination : '');
     const getOffer = (point) => this.#offers.find((item) => item.type === point.type ? item.type === point.type : '');
 
-    if (points.length <= 0) {
+    if (!points.length) {
       this.#renderMessage();
       return;
     }
-    this.#renderSort(TYPES_OF_SORT);
-    this.#renderList();
+
     for (const point of points) {
-      this.#renderPoint(point, getDestination(point), getOffer(randomPoint()));
+      new PointPresenter(this.#eventsListElement).init(point, getDestination(point), getOffer(point));
     }
   }
 
@@ -111,8 +64,9 @@ export default class EventsPresenter {
     render(eventsSortElement, this.#eventsElement);
   }
 
-  #renderList() {
+  #renderList(points) {
     render(this.#eventsListElement, this.#eventsElement);
+    this.#renderPoints(points);
   }
 
   #renderInfo() {
